Update icon/fix issue with DSI Bootsplash not working...
Fixed issue with DSi Boot Splash not showing if Hiya Logo is disabled.

Note that preivous commit is not compatible with this. It has been
reverted. Please re-do that code. (or wait for NoCash to change Unlaunch
so it doesn't set the skip DSi splash flag for bootcode.dsi payloads)

Also updated readme.md and icon.bmp.